% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Print the sounds animals make}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{A string - the name of the animal}

\item{sound}{A string - the name of the sound}
}
\value{
A string - 'The \{animal\} says \{sound\}!'
}
\description{
Print the sounds animals make
}
\examples{
animal_sounds('dog','woof')
}
